Move client and query charts into widgets
